Updated the instructions box, added instructions to many tests, updated reaction test, hooked reaction test up to grapher, some small UI changes
I did a bit of everything in this one.

I updated the look of the “help” dialogue box to something a bit
cleaner. I added the instruction box to all the tests but the ball
test, since that one has the dialogue boxes to guide the user.

I also finished up the reaction test by putting instructions, and
having the test start after the first click, so the timer doesn’t start
when the user enters the screen. I also had it display the results, and
kick the user out of the test when finished.

Finally, I set up a graph demo, and hooked the data from the reaction
test into it.